동행 style img, age, gender commit

+++ gasil/src/main/java/org/joonzis/domain/UserAuthVO.java
@@ -11,6 +11,6 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class UserAuthVO {
 	private String user_id, user_pw, user_email, user_name, user_birth, user_phone, gender, user_nick, user_pic, auth, grade;
-	private int bno;
+	private long bno;
 	private Date joindate;
 }
